+++ src/calendar/index.less
@@ -5,12 +5,8 @@
   flex-direction: column;
   height: 80vh;
 
-  &__header {
-    flex-shrink: 0;
-    box-shadow: 0 2px 10px rgba(125, 126, 128, .16);
-  }
-
-  &__title,
+  &-header__title,
+  &-header__month,
   &__month-title {
     height: 44px;
     font-weight: @font-weight-bold;
@@ -18,23 +14,32 @@
     text-align: center;
   }
 
-  &__title {
-    font-size: @font-size-lg;
+  &-header {
+    flex-shrink: 0;
+    box-shadow: 0 2px 10px rgba(125, 126, 128, 0.16);
+
+    &__title {
+      font-size: @font-size-lg;
+    }
+
+    &__month {
+      font-size: @font-size-md;
+    }
+
+    &__weekdays {
+      display: flex;
+    }
+
+    &__weekday {
+      flex: 1;
+      font-size: @font-size-sm;
+      line-height: 30px;
+      text-align: center;
+    }
   }
 
   &__month-title {
     font-size: @font-size-md;
-  }
-
-  &__weekdays {
-    display: flex;
-  }
-
-  &__weekday {
-    flex: 1;
-    font-size: @font-size-sm;
-    line-height: 30px;
-    text-align: center;
   }
 
   &__body {
