+++ src/calendar/utils.ts
@@ -1,3 +1,13 @@
+import { createNamespace } from '../utils';
+
+const [createComponent, bem, t] = createNamespace('calendar');
+
+export { createComponent, bem, t };
+
+export function formatMonthTitle(date: Date) {
+  return t('monthTitle', date.getFullYear(), date.getMonth() + 1);
+}
+
 export function compareMonth(date1: Date, date2: Date) {
   const year1 = date1.getFullYear();
   const year2 = date2.getFullYear();
